requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
pydantic>=2.7
pyyaml>=6.0.1
orjson>=3.10
tqdm>=4.66
pandas>=2.2
bert-score>=0.3.13
huggingface-hub>=0.23
jsonschema>=4.22
rich>=13.7
python-dotenv>=1.0
datasets>=2.19
python-toon